refactor: extract Content component and improve interceptor timeout
