+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-nitro-sse",
-  "version": "1.5.0",
+  "version": "1.5.1",
   "description": "High-performance Server-Sent Events (SSE) for React Native, powered by Nitro Modules.",
   "main": "./lib/module/index.js",
   "types": "./lib/typescript/src/index.d.ts",
